--- verifier/src/test/java/com/aws/jverify/verifier/tests/Allocate.java
@@ -1,38 +1,0 @@
-package com.aws.jverify.verifier.tests;
-
-import com.aws.jverify.testengine.JVerifyTest;
-
-import com.aws.jverify.*;
-import static com.aws.jverify.JVerify.*;
-
-// Class that test the support of array allocation and accesses
-@JVerifyTest(dafnyVerified = 5, dafnyErrors = 0)
-class Allocate {
-    public static IntPair allocateInReturn(int a, int b) {
-        postcondition((IntPair p) -> fresh(p) &&  p.getA() == a);
-        return new IntPair(a, b);
-    }
-
-    public static void test(int a) {
-        IntPair p = allocateInReturn(a,2);
-        check(p.getA() == a);
-    }
-}
-
-class IntPair {
-    private int a;
-    private int b;
-
-    public IntPair(int a_, int b_) {
-        postcondition(this.a == a_);
-        postcondition(this.b == b_);
-        this.a = a_;
-        this.b = b_;
-    }
-
-    @Pure
-    public int getA() {
-        reads(this);
-        return a;
-    }
-}
