--- verifier/src/test/java/com/aws/jverify/verifier/tests/AssertFalse.java
@@ -1,13 +1,0 @@
-package com.aws.jverify.verifier.tests;
-
-import com.aws.jverify.testengine.JVerifyTest;
-
-import static com.aws.jverify.JVerify.*;
-
-@JVerifyTest(exitCode = 4, dafnyVerified = 0, dafnyErrors = 1)
-class AssertFalse {
-    static void Foo() {
-        check(false);
-//      ^^^^^^^^^^^^ Error: assertion might not hold
-    }
-}
