--- verifier/src/test/java/com/aws/jverify/verifier/tests/ConstructorsErrors.java
@@ -1,19 +1,0 @@
-package com.aws.jverify.verifier.tests;
-
-import com.aws.jverify.testengine.JVerifyTest;
-
-import static com.aws.jverify.JVerify.postcondition;
-
-@JVerifyTest(exitCode = 2)
-public class ConstructorsErrors {
-    class Box {
-        private int value;
-
-        public Box(int value_) {
-            this.value = value_;
-            postcondition(this.value == value_);
-//                       ^ error: for constructors, calls to JVerify contract methods must come either immediately after the call to 'super' or 'this', or at the end of the body
-            var b = 3;
-        }
-    }
-}
