+++ verifier/src/test/java/com/aws/jverify/verifier/tests/classes/ClassesExtendingClassesErrors.java
@@ -1,0 +1,27 @@
+package com.aws.jverify.verifier.tests.classes;
+
+import com.aws.jverify.Pure;
+import com.aws.jverify.testengine.JVerifyTest;
+
+@JVerifyTest(exitCode = 22)
+public class ClassesExtendingClassesErrors {}
+
+class ExtenderErrors extends ExtendeeErrors {
+    
+    @Pure
+    @Override
+    public int pureVirtual(int input) {
+//             ^^^^^^^^^^^ Error: fully defined function 'pureVirtual' is inherited from trait 'ExtendeeErrors' and is not allowed to be re-declared
+ 
+        return 3;
+    }
+}
+
+
+abstract class ExtendeeErrors {
+    
+    @Pure
+    public int pureVirtual(int input) {
+        return 3;
+    }
+}
