+++ verifier/src/test/java/com/aws/jverify/verifier/tests/records/ResolutionErrorsStringMethods.java
@@ -1,0 +1,14 @@
+package com.aws.jverify.verifier.tests.records;
+
+import com.aws.jverify.testengine.JVerifyTest;
+
+import static com.aws.jverify.JVerify.*;
+
+@JVerifyTest(exitCode = 2)
+class ResolutionErrorsStringMethods {
+    static void stringFormatted() {
+        check("hello %s".formatted("world").length() == 11);
+//                                ^ error: String method formatted(java.lang.Object...) is not supported
+//                                 ^ error: JCNewArray in an expression is not supported
+    }
+}
